% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMonacoEditor.R
\name{shinyMonacoEditor}
\alias{shinyMonacoEditor}
\title{Monaco Editor in Shiny}
\usage{
shinyMonacoEditor(indentSize = 2)
}
\arguments{
\item{indentSize}{an integer, at least 2, the number of spaces of the
indentation}
}
\description{
Launch a Shiny app with the Monaco editor.
}
\examples{
if(interactive()){
  shinyMonacoEditor()
}
}
