% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMonacoEditor.R
\name{shinyMonacoEditor}
\alias{shinyMonacoEditor}
\title{Monaco Editor in Shiny}
\usage{
shinyMonacoEditor()
}
\description{
Launch a Shiny app with the Monaco editor.
}
\examples{
if(interactive()){
  shinyMonacoEditor()
}
}
